Tell mysql++ where to find libmysqlclient (apparently required on Mac)
Explicitly tell OpenSSL to build for 64-bit if on Mac 64-bit. By default it builds a 32-bit library even if on Mac 64-bit. Not only that, it acknowledges it's Mac 64-bit and tells you 'if you want to build for 64-bit, you will have to force me to do so'. Pointless.