WT-11270: Revert Haru to 2.3.0
The 32bit Windows binary for Haru (HPDF) doesn't seem to function
correctly when build as part of the Windows CMake build.

As a result the bundled `hpdf(d).lib` files and `hpdf(d).dll` files
result in link errors when used.

It seems this error was not yet present with any builds before Haru
2.3.0. This version will be bumped again later, when the correct 32bit
binary can be generated.