Update openSSL dependency version
Update openSSL dependency version from 1.1.1 to 3.4.0. This implies a
change in the way the patch is done the name of the library compiled
with the debug option: it now uses a custom configuration instead of
updating the build.info file. Another small change is that now the
license of openSSL is named LICENSE.txt instead of LICENSE.